Refactor => Merged all routes into 1 router => deleted add photo endpoint => deleted images folder
