--- Routes/delete_book.js
@@ -1,14 +1,0 @@
-//Controllers
-const delete_book_controller = require("../controllers/DeleteBookController");
-
-//External
-const express = require("express");
-
-//Config
-const router = express.Router();
-
-//route
-router.delete("/delete_book", delete_book_controller.delete_book);
-
-
-module.exports = router;
