+++ app.js
@@ -1,14 +1,9 @@
 //Core node
 const path = require("path");
 
-//routers
-const add_book_router = require("./Routes/add_book")
-const delete_book_router = require("./Routes/delete_book")
-const search_for_book = require("./Routes/search_for_book")
-const worth_buying = require("./Routes/worth_buying")
-const fetch_books = require("./Routes/fetch_books")
-const upload_photo = require("./Routes/upload_photo")
-const get_conditions = require("./Routes/get_conditions")
+//router
+const router = require("./Routes/router")
+
 
 //External
 const bodyParser = require("body-parser");
@@ -58,15 +53,7 @@ server.use("./images", express.static(path.join(__dirname, "images")))
 
 server.use(bodyParser.urlencoded({ extended: false }));//Set up the body parser
 
-
-server.use(add_book_router)
-server.use(delete_book_router)
-server.use(search_for_book)
-server.use(worth_buying)
-server.use(fetch_books)
-server.use(upload_photo)
-server.use(get_conditions)
-
+server.use(router)
 
 //* Database connection
